pyproject.toml
[project]
name = "wikipedia-flashcards-agent"
version = "0.1.0"
description = "A Wikipedia flashcard generator using Pydantic AI agents"
requires-python = ">=3.12"
dependencies = [
    "pydantic-ai>=0.0.1",
    "openai>=1.0.0",
    "wikipedia>=1.4.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
flashcards = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "models"]
